pyproject.toml
[tool.poetry]
name = "matchmaker"
version = "0.1.0"
description = ""
authors = ["Zach Bellay <zachbellay@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
psycopg2 = "^2.9.5"
python-dotenv = "^1.0.0"
pgvector = "^0.1.6"
django-ninja = "^0.21.0"
celery = "5.2.7"
django-celery = "3.1.17"
django-celery-beat = "2.4.0"
sentence-transformers = "^2.2.2"
django-extensions = "^3.2.1"
httpx = "^0.23.3"
beautifulsoup4 = "^4.12.0"
mmh3 = "^3.1.0"
django-vite = "^2.0.2"
whitenoise = "^6.4.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.260"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
